+++ tests/cases/timer_one_shot/service.cpp
@@ -30,7 +30,7 @@ void skull_service_init(skullcpp::Service& service, const skull_config_t* config
     skullcpp::Config::instance().load(config);
 
     // Create a timer job (should be failed)
-    int ret = service.createJob(1000, skull_BindSvcJob(_timerjob));
+    int ret = service.createJob((uint32_t)1000, skull_BindSvcJob(_timerjob));
     if (!ret) {
         SKULLCPP_LOG_INFO("init", "service job create successful");
     } else {
